# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IntelDoc API",
    "description": "IntelDoc backend — Эндокор pilot MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/patient/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Patient Otp Request",
        "operationId": "patient_otp_request_v1_auth_patient_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/patient/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Patient Otp Verify",
        "operationId": "patient_otp_verify_v1_auth_patient_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/web/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Web Login",
        "operationId": "web_login_v1_auth_web_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Session",
        "operationId": "session_v1_auth_session_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/partner-context": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Partner Context",
        "operationId": "partner_context_v1_onboarding_partner_context_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "enc",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerContextOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/account-draft": {
      "put": {
        "tags": [
          "onboarding"
        ],
        "summary": "Account Draft",
        "operationId": "account_draft_v1_onboarding_account_draft_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDraftIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "type": "object",
                  "title": "Response Account Draft V1 Onboarding Account Draft Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/commit": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Commit",
        "operationId": "commit_v1_onboarding_commit_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/sms-otp/request": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Sms Otp Request",
        "operationId": "sms_otp_request_v1_onboarding_sms_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/sms-otp/verify": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Sms Otp Verify",
        "operationId": "sms_otp_verify_v1_onboarding_sms_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Sms Otp Verify V1 Onboarding Sms Otp Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountDraftIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dob": {
            "type": "string",
            "title": "Dob"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "name",
          "dob",
          "phone"
        ],
        "title": "AccountDraftIn"
      },
      "AckMechanism": {
        "type": "string",
        "enum": [
          "scroll_to_end",
          "a11y_checkbox",
          "direct_tick",
          "not_applicable"
        ],
        "title": "AckMechanism"
      },
...
      "CommitIn": {
        "properties": {
          "department_public_id": {
            "type": "string",
            "format": "uuid",
            "title": "Department Public Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "dob": {
            "type": "string",
            "title": "Dob"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "oms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oms"
          },
          "snils": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snils"
          },
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentRecordIn"
            },
            "type": "array",
            "title": "Consents"
          },
          "document_hash": {
            "type": "string",
            "title": "Document Hash"
          }
        },
        "type": "object",
        "required": [
          "department_public_id",
          "name",
          "dob",
          "gender",
          "phone",
          "document_hash"
        ],
        "title": "CommitIn"
      },
      "CommitOut": {
        "properties": {
          "patient_public_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Public Id"
          },
          "grant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GrantOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated"
          }
        },
        "type": "object",
        "required": [
          "patient_public_id",
          "grant",
          "deduplicated"
        ],
        "title": "CommitOut"
      },
...
      "ConsentRecordIn": {
        "properties": {
          "consent_type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "legal_text_version": {
            "type": "string",
            "title": "Legal Text Version"
          },
          "ack_mechanism": {
            "$ref": "#/components/schemas/AckMechanism"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "default": true
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "sms_confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sms Confirmed"
          }
        },
        "type": "object",
        "required": [
          "consent_type",
          "legal_text_version",
          "ack_mechanism"
        ],
        "title": "ConsentRecordIn"
      },
      "ConsentType": {
        "type": "string",
        "enum": [
          "pdn_general",
          "pdn_special",
          "medical_data",
          "clinic_transfer",
          "ocr_ai",
          "cross_border",
          "marketing"
        ],
        "title": "ConsentType"
      },
...
      "GrantOut": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "granted_to_type": {
            "type": "string",
            "title": "Granted To Type"
          },
          "data_scope": {
            "type": "string",
            "title": "Data Scope"
          },
          "valid_from": {
            "type": "string",
            "title": "Valid From"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "last_viewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Viewed At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "granted_to_type",
          "data_scope",
          "valid_from",
          "status"
        ],
        "title": "GrantOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogoutIn": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "LogoutIn"
      },
...
      "OtpRequestIn": {
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 5,
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "OtpRequestIn"
      },
      "OtpVerifyIn": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "OtpVerifyIn"
      },
      "PartnerContextOut": {
        "properties": {
          "partner_public_id": {
            "type": "string",
            "format": "uuid",
            "title": "Partner Public Id"
          },
          "partner_name": {
            "type": "string",
            "title": "Partner Name"
          },
          "partner_short_name": {
            "type": "string",
            "title": "Partner Short Name"
          },
          "department_public_id": {
            "type": "string",
            "format": "uuid",
            "title": "Department Public Id"
          },
          "department_name": {
            "type": "string",
            "title": "Department Name"
          }
        },
        "type": "object",
        "required": [
          "partner_public_id",
          "partner_name",
          "partner_short_name",
          "department_public_id",
          "department_name"
        ],
        "title": "PartnerContextOut"
      },
      "RefreshIn": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn"
      },
...
      "SessionOut": {
        "properties": {
          "subject_public_id": {
            "type": "string",
            "title": "Subject Public Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "partner_id": {
            "type": "string",
            "title": "Partner Id"
          },
          "clinic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinic Id"
          }
        },
        "type": "object",
        "required": [
          "subject_public_id",
          "role",
          "partner_id"
        ],
        "title": "SessionOut"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "subject_public_id": {
            "type": "string",
            "title": "Subject Public Id"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "role",
          "subject_public_id"
        ],
        "title": "TokenOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebLoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "WebLoginIn"
      }
    }
  }
}